+++ angular-workspace/projects/trimble-oss/modus-angular-components/src/lib/stencil-generated/index.ts
@@ -5,6 +5,7 @@ export const DIRECTIVES = [
   d.ModusAccordion,
   d.ModusAccordionItem,
   d.ModusAlert,
+  d.ModusAutocomplete,
   d.ModusBadge,
   d.ModusBreadcrumb,
   d.ModusButton,