+++ react-workspace/src/components/stencil-generated/index.ts
@@ -11,6 +11,7 @@ defineCustomElements();
 export const ModusAccordion = /*@__PURE__*/createReactComponent<JSX.ModusAccordion, HTMLModusAccordionElement>('modus-accordion');
 export const ModusAccordionItem = /*@__PURE__*/createReactComponent<JSX.ModusAccordionItem, HTMLModusAccordionItemElement>('modus-accordion-item');
 export const ModusAlert = /*@__PURE__*/createReactComponent<JSX.ModusAlert, HTMLModusAlertElement>('modus-alert');
+export const ModusAutocomplete = /*@__PURE__*/createReactComponent<JSX.ModusAutocomplete, HTMLModusAutocompleteElement>('modus-autocomplete');
 export const ModusBadge = /*@__PURE__*/createReactComponent<JSX.ModusBadge, HTMLModusBadgeElement>('modus-badge');
 export const ModusBreadcrumb = /*@__PURE__*/createReactComponent<JSX.ModusBreadcrumb, HTMLModusBreadcrumbElement>('modus-breadcrumb');
 export const ModusButton = /*@__PURE__*/createReactComponent<JSX.ModusButton, HTMLModusButtonElement>('modus-button');