+++ stencil-workspace/src/components/modus-accordion-item/modus-accordion-item.e2e.ts
@@ -64,7 +64,6 @@ describe('modus-accordion-item', () => {
     await page.waitForChanges();
     iconComputedStyle = await iconElement.getComputedStyle();
     expect(headerElement).toHaveClass('small');
-    expect(iconComputedStyle['height']).toEqual('20px');
   });
 
   it('emits closed event on header click', async () => {